增加DataManagerUtil类设置内存数据保护对应函数
1. 增加crypto中crc32返回值类型为unsigned long，并且修改crc32.h头文件，增加extern C修饰，以支持cpp调用

+++ Classes/comm/Util/crypto.cpp
@@ -35,7 +35,7 @@ std::string bailin::util::crypto::getXORString( const char* key, const std::stri
 	return retStr;
 }
 
-int bailin::util::crypto::Crc32(const void *buf, size_t bufLen)
+unsigned long bailin::util::crypto::Crc32(const void *buf, size_t bufLen)
 {
     return Crc32_ComputeBuf(0, buf, bufLen);
 }
